Change user, role details query to work for pending/rejected provider
In the OIDC flow, a user may have registered for a user profile, but may not
have any approved roles associated with them, e.g. a user with provider role
in pending/rejected status. The old user query returned 0 rows if such a user
came thru the OIDC flow. The user would then get a 404 in the List UserProfile
 API, and a 409 Conflict in the Create UserProfile API.

Now, the query returns 1 row with an empty roles array if the user has
no approved roles. Thus a user with an empty roles array in the List UserProfile
API can be assumed to be a user with provider role in pending/rejected status.