Bugfix - make orgId always required for adding delegate role
Previously, orgId was only required when the user had an
approved Consumer role and was requesting to be a delegate. This did not
consider that the user may have had a pending/rejected provider role. To
simplify this flow, we now always need orgId if delegate role
has been requested and overwrite the orgId in the DB in case it was already there.