Update KC AuthN token size limit, update introspect token size, add Authorization header to notification APIs
- Keycloak requires to add redirect_uris for each domain/endpoint that would use a client.
CORS also needs to be configured for each of the domains/clients. But wildcard sub-domains
cannot be configured for both, hence each new subdomain needs to be added. The Keycloak JWT
(AuthN) token contains an 'allowed-origins' field that has all the CORS values in it. Since
every cities' catalogue is a separate sub-domain, 'allowed-origin' would contain each subdomain.
The Keycloak JWT token size is increased to allow these bigger tokens

- Increased introspect token size to 640 as some tokens were generated which were greater than 512

- Added Authorization header to notif APIs. The token check was enforced even without this, but added
to have uniformity