Add integ tests for cli-sec regen, more for notif API + bugfix - Added tests for client secret regen API - Added additional tests for notification API, to see if tokens can be obtained after approve notification
- Updated OpenAPI spec for PUT /user/profile API

The PUT /user/profile API uses the `oneOf` to restrict users to either
(roles,orgId) schema or (clientId) schema in the request (role add **or** cli-sec regen).
So an object with string `clientId` and array of string `roles` will not be
allowed because only either is allowed by `oneOf`.

However, an object with string `clientId` and integer/bool/string `roles`
was being allowed. This is because the spec did not care about additional key-values
in the object that are not specified by any schema and just passes them through.
`roles` as an array was specified in the (roles,orgId) schema, but `roles` as int/string
etc. did not trigger any validation exception. The UpdateProfile request object
class expects `roles` to be an array, so an internal error was caused.

So added `additionalProperties:false` to the individual schemas for that API.
This strictly forbids additional keys that do not appear in the schema.