Email Service Implementation The PR contains the implementation of an email service that allows the provider and delegate to receive an email whenever a consumer requests to access a resource.
Files Modified or Created:
config-file: Added config parameters that contain all the information required to create a connection to send an email.
EmailInfo: This class contains all the data required to create an email, such as information related to the body of the email, sender's email ID, and receiver's email ID.
EmailClient: This class does all the work, from extracting data from the EmailInfo object required to create the body of the email to sending the email to the provider and delegate of the resource requested.
PolicyServiceImpl: Created an object emailInfo and set all its parameters in the createNotification method. After that, called the sendEmail method of the EmailClient class to send emails to the right person.
PolicyVerticle: Creation of EmailClient object and getting the config required to send an email from the config file is done in this class.
Constant: Constants such as a Postgres query to get all the delegate IDs for a given provider ID and the raw email body are added in this file.